RM update + db fix

+++ src/SciVacancies.PostgresqlDB/up/001_22_autoincremented_fields.sql
@@ -8,9 +8,9 @@ SET client_min_messages = warning;
 SET search_path = public, pg_catalog;
 
 CREATE SEQUENCE org_vacancy_id_seq;
-ALTER TABLE "org_vacancies" ADD COLUMN "id" integer NOT NULL default nextval('org_vacancy_id_seq');
+ALTER TABLE "org_vacancies" ADD COLUMN "id" bigint NOT NULL default nextval('org_vacancy_id_seq');
 ALTER SEQUENCE org_vacancy_id_seq owned by org_vacancies.id;
 
 CREATE SEQUENCE res_vacancyapplications_id_seq;
-ALTER TABLE "res_vacancyapplications" ADD COLUMN "id" integer NOT NULL default nextval('res_vacancyapplications_id_seq');
+ALTER TABLE "res_vacancyapplications" ADD COLUMN "id" bigint NOT NULL default nextval('res_vacancyapplications_id_seq');
 ALTER SEQUENCE res_vacancyapplications_id_seq owned by res_vacancyapplications.id;
